+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Internal Changes
 * Refactored `setup.py` for better packaging and improved long description through README.md ([#57](https://github.com/zscaler/zscaler-sdk-python/pull/57))
-
+* Refactored Integration Tests by removing `async` decorators ([#63](https://github.com/zscaler/zscaler-sdk-python/pull/63))
 
 
 ## 0.1.0 (April, 18 2024)